DD-30682 progression.event.case-ejected listener fix. DD-30691 progression.event.hearing-marked-as-duplicate-for-case listener fix.

+++ progression-event/progression-event-listener/src/main/java/uk/gov/moj/cpp/prosecutioncase/event/listener/HearingMarkedAsDuplicateEventListener.java
@@ -68,11 +68,13 @@ public class HearingMarkedAsDuplicateEventListener {
             LOGGER.debug("received event progression.event.hearing-marked-as-duplicate-for-case hearingId: {} caseId:{}", hearingId, caseId);
         }
 
-        defendantIds.forEach(defendantId -> {
-            caseDefendantHearingRepository.removeByHearingIdAndCaseIdAndDefendantId(hearingId, caseId, defendantId);
-            matchDefendantCaseHearingRepository.removeByHearingIdAndCaseIdAndDefendantId(hearingId, caseId, defendantId);
-        });
+        final HearingEntity hearingEntity = hearingRepository.findBy(hearingId);
+        if (nonNull(hearingEntity)) {
+            defendantIds.forEach(defendantId -> {
+                caseDefendantHearingRepository.removeByHearingIdAndCaseIdAndDefendantId(hearingId, caseId, defendantId);
+                matchDefendantCaseHearingRepository.removeByHearingIdAndCaseIdAndDefendantId(hearingId, caseId, defendantId);
+            });
+        }
     }
-
 
 }
